implement again flow and adding field for model

+++ src/services/deviceService.js
@@ -1,9 +1,9 @@
 import { DEVICE_REPOSITORY } from '#repositories/deviceRepository.js'
 import ApiError from '#utils/ApiError.js'
 import { ERROR_CODES } from '#constants/errorCode.js'
-import { USER_REPOSITORY } from '#repositories/userRepository.js'
 import { escapeRegex } from '#utils/formatterUtil.js'
 import { mapMongoosePagination } from '#utils/pagination.js'
+import { USER_SERVICE } from '#services/userService.js'
 
 const getDeviceById = async (deviceId) => {
   const device = await DEVICE_REPOSITORY.getDeviceById(deviceId)
@@ -12,8 +12,7 @@ const getDeviceById = async (deviceId) => {
 }
 
 const createDevice = async (data, createdBy = null) => {
-  const user = await USER_REPOSITORY.getUserById(createdBy)
-  if (!user) throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Người tạo không tồn tại'])
+  const user = await USER_SERVICE.getUserById(createdBy)
 
   const { name, type, brand, model } = data
   const existingDevice = await DEVICE_REPOSITORY.getDeviceByName(name)
@@ -33,8 +32,7 @@ const updateDevice = async (deviceId, data, updatedBy = null) => {
 
   const device = await DEVICE_REPOSITORY.getDeviceById(deviceId)
   if (!device) throw new ApiError(ERROR_CODES.NOT_FOUND, ['Thiết bị không tồn tại'])
-  const user = await USER_REPOSITORY.getUserById(updatedBy)
-  if (!user) throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Người cập nhật không tồn tại'])
+  const user = await USER_SERVICE.getUserById(updatedBy)
   const updatedDevice = {}
   if (name && name !== device.name) {
     const existingDevice = await DEVICE_REPOSITORY.getDeviceByName(name)
@@ -52,8 +50,7 @@ const updateDevice = async (deviceId, data, updatedBy = null) => {
 const updateDeviceStatus = async (deviceId, isActive, updatedBy = null) => {
   const device = await DEVICE_REPOSITORY.getDeviceById(deviceId)
   if (!device) throw new ApiError(ERROR_CODES.NOT_FOUND, ['Thiết bị không tồn tại'])
-  const user = await USER_REPOSITORY.getUserById(updatedBy)
-  if (!user) throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Người cập nhật không tồn tại'])
+  const user = await USER_SERVICE.getUserById(updatedBy)
 
   if (typeof isActive !== 'boolean') {
     throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Trạng thái thiết bị không hợp lệ'])
@@ -68,6 +65,8 @@ const updateDeviceStatus = async (deviceId, isActive, updatedBy = null) => {
 const deleteDevice = async (deviceId) => {
   const device = await DEVICE_REPOSITORY.getDeviceById(deviceId)
   if (!device) throw new ApiError(ERROR_CODES.NOT_FOUND, ['Thiết bị không tồn tại'])
+  if (!device.isActive) throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Chỉ có thể xoá thiết bị đang hoạt động'])
+  if (device.isDeleted) throw new ApiError(ERROR_CODES.BAD_REQUEST, ['Thiết bị đã bị xoá'])
   return DEVICE_REPOSITORY.deleteDeviceById(deviceId)
 }
